Update docs to rand 0.6.5
I want to preface this by saying that I'm totally new to Rust, so please correct any mistakes I might have done, if you feel like this pull request is useful.

Version 0.3.14 of the rand crate was released way back in 2016, so I thought it would be nice to have a more up-to-date version to play around with.

I also reworded a section about `rand` depending on `libc` since that doesn't appear to be the case anymore for Windows at least. Plus, `rand` also has a lot more dependencies now.

I built it and tested it on Windows 10, but would like if someone did it on Linux or something to make the `cargo` build output representative of that. I also couldn't get the output to say that it was updating the "registry", even after deleting the cargo registry on my PC. All I got was `Updating crates.io index`. Does anyone know why?